requirements.txt
fake-bpy-module==20250630
